+++ src/fetch.rs
@@ -128,6 +128,22 @@ impl FetchData {
 					}
 				},
 			},
+			FetchType::Shell => FetchData {
+				label: "shell",
+				icon: "",
+				text: {
+					match env::var("SHELL") {
+						Ok(shell) => match std::path::Path::new(&shell).file_name() {
+							Some(shell) => match shell.to_str() {
+								Some(shell) => shell.to_string(),
+								None => "unknown".to_string(),
+							},
+							None => "unknown".to_string(),
+						},
+						Err(_) => "unknown".to_string(),
+					}
+				},
+			},
 		}
 	}