+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to
   run.
 - Highly improved error handling - program should no longer panic other than in
   catastrophic cases, thanks to implementing the `anyhow` crate
+- New data option: `shell`
 
 ### Changed